in6 se inicio la modificacion del frontEnd y se organizaron las carpetas existentes de este
