--- react-front/src/assets/adminStyle.css
@@ -1,76 +1,0 @@
-:root{
-    --principal-color: #1d40af;
-}
-
-.admin-menu{
-    position: fixed;
-    font-size: 25px;
-    z-index: 3;
-    top:300px;
-    left: 15px;
-    transform: rotate(-90deg);
-    border-radius: 5px;
-    color: var(--principal-color);
-}
-
-.admin-menu:hover{
-    transform: rotate(0deg);
-    cursor: pointer;
-}
-
-.container-admin{
-    display: flex;
-    flex-direction: column;
-    position: fixed;
-    background-color: var(--principal-color);
-    top: 330px;
-    width: 370px;
-    height: 250px;
-    z-index: 3;
-    left: 15px;
-    box-shadow: 2px 2px 10px 2px black;
-    border: 2px rgb(255, 255, 255) solid;
-    border-radius: 12px;
-}
-
-.menu-class-close{
-    box-sizing: content-box;
-    height: 25px;
-    width: 95%;
-    display: flex;
-    justify-content: flex-end;
-    margin-top: 7px;
-}
-
-.menu-close:hover{
-    cursor: pointer;
-    text-shadow: 0 5px 3px black;
-}
-
-.li-admin{
-    display: flex;
-    font-size: 18px;
-    color: white;
-}
-
-.li-admin:hover{
-    cursor: pointer;
-    text-shadow: 0 8px 2px black;
-}
-
-.fa-mar{
-    margin-right: 20px;
-}
-
-.ul-admin{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    height: 98%;
-}
-
-@media (max-width:386px) {
-    .container-admin{
-        width: 90%;
-    }
-}
