+++ react-front/src/components/ModalConsult.jsx
@@ -2,8 +2,6 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import { Modal } from "react-bootstrap";
 import DataTable from "react-data-table-component";
-import Swal from "sweetalert2";
-import ModalUpdate from "./ModalUpdate";
 
 const ModalConsult = ({ lgShowConsult, closeModalConsult }) => {
     const [usersDb, setUsersDb] = useState([]);
@@ -30,11 +28,11 @@ const ModalConsult = ({ lgShowConsult, closeModalConsult }) => {
             sortable: true
         },
         {
-            name: "Acciones",
-            selector: row => (
+            name:"Acciones",
+            selector: row =>(
                 <div>
-                    <i id={row.id} onClick={dataDelete} class="fa-solid fa-trash btn btn-danger" style={{ fontSize: 12, marginRight: "2px" }}></i>
-                    <i id={row.id} onClick={showModal} class="fa-solid fa-pen btn btn-primary" style={{ fontSize: 12, marginRight: "2px" }}></i>
+                    <i class="fa-solid fa-trash btn btn-primary" style={{fontSize:16, marginRight:"2px"}}></i>
+                    <i class="fa-solid fa-pen btn btn-primary" style={{fontSize:16,marginRight:"2px"}}></i>
                 </div>
             )
         }
@@ -46,68 +44,9 @@ const ModalConsult = ({ lgShowConsult, closeModalConsult }) => {
                 setUsersDb(response.data);
             })
             .catch(err => {
-                console.log("error", err)
+                console.log(["error antes de controlador", err, 500]);
             });
     }, []);
-
-    const updateUser = () => {
-        axios.get(`${URI}userConsult`)
-            .then(response => {
-                setUsersDb(response.data);
-            })
-            .catch(err => {
-                console.log("error", err)
-            });
-    }
-
-    const [lgShow, setLgConsult] = useState(false);
-    const [userID, setUserID] = useState(0);
-
-    const showModal = (e) => {
-        setUserID(e.target.id);
-        setLgConsult(true);
-    }
-
-    const closeModal = () => {
-        setLgConsult(false)
-        updateUser();
-    }
-
-
-    const dataDelete = (e) => {
-        Swal.fire({
-            title: "Estás seguro de eliminar un usuario?",
-            text: "Eliminaras al usuario con el id " + e.target.id,
-            icon: "warning",
-            showCancelButton: true,
-            confirmButtonColor: "#3085d6",
-            cancelButtonColor: "#d33",
-            confirmButtonText: "Confirmar",
-            cancelButtonText: "Cancelar"
-        }).then((result) => {
-            if (result.isConfirmed) {
-                axios.delete(`${URI}userDelete/${e.target.id}`)
-                    .then(response => {
-                        console.log("Usuario eliminado", response.data)
-                        axios.get(`${URI}userConsult`)
-                            .then(response => {
-                                setUsersDb(response.data);
-                            })
-                            .catch(err => {
-                                console.log("Error al obtener usuarios", err);
-                            });
-                        Swal.fire({
-                            title: "Exito!",
-                            text: "El usuario fue eliminado",
-                            icon: "success"
-                        });
-                    })
-                    .catch(error => {
-                        console.log("Error eliminando", error)
-                    })
-            }
-        });
-    }
 
     return (
         <>
@@ -115,22 +54,23 @@ const ModalConsult = ({ lgShowConsult, closeModalConsult }) => {
                 size="xl"
                 show={lgShowConsult}
                 onHide={closeModalConsult}
-                aria-labelledby="example-modal-sizes-title-xl"
+                aria-labelledby="example-modal-sizes-title-lg"
             >
                 <Modal.Header closeButton style={{ fontSize: 13 }}>
-                    <Modal.Title style={{ fontSize: 20 }} id="example-modal-sizes-title-xl">
+                    <Modal.Title style={{ fontSize: 20 }} id="example-modal-sizes-title-lg">
                         Usuarios
                     </Modal.Title>
                 </Modal.Header>
                 <Modal.Body>
-                    <DataTable columns={columns}
-                        data={usersDb}
+                    <div>
+                        <DataTable columns={columns} 
+                        data={usersDb}                 
                         pagination
                         highlightOnHover
-                        pointerOnHover />
+                        pointerOnHover/>
+                    </div>
                 </Modal.Body>
             </Modal>
-            <ModalUpdate lgShow={lgShow} closeModal={closeModal} userID={userID} updateUser={updateUser} />
         </>
     )
 }
