--- react-front/src/pages/IniciarSesion.js
@@ -1,5 +1,0 @@
-const iniciarSesion = ()=>{
-    return <h1>Inicio de sesión</h1>
-}
-
-export default iniciarSesion;
