+++ react-front/src/pages/InisiarSesion/IniciarSesion.js
@@ -1,0 +1,5 @@
+const iniciarSesion = ()=>{
+    return <h1>Inicio de sesión</h1>
+}
+
+export default iniciarSesion;
