--- react-front/src/pages/Registrarse.js
@@ -1,3 +1,0 @@
-const Registrarse = ()=>{
-    return <h1>En esta página se hace el registro</h1>
-}
