--- react-front/src/pages/Reparacion.js
@@ -1,66 +1,0 @@
-const Reparacion = () => {
-	return <div>
-		<section className="reparacionrs">
-			<div className="informacionrs">
-				<div className="titulors">
-					<h2>SERVICIO TÉCNICO, REPARACIÓN, ARREGLO Y MANTENIMIENTO DE LAVADORAS A
-						DOMICILIO</h2>
-				</div>
-				<div className="contenedorrs">
-					<div className="contenedor2rs">
-						<div>
-							<h2>Reparación y Servicio Técnico de Lavadoras Rapido, Efectivo y
-								Economico</h2>
-							<p>Lavaexpress es una empresa multimarca independiente, experta en
-								reparación, soporte técnico y mantenimiento de lavadoras y
-								secadoras
-								del hogar. <br />
-								El servicio que ofrecemos es con técnicos profesionales y
-								certificados para
-								brindar soluciones a cualquier reparación o servicio que requieras
-								con
-								resultados garantizados y de alta calidad.
-								<br />
-								Nuestros técnicos están calificados para realizar reparaciones y
-								dar
-								servicio de
-								mantenimiento a diversas marcas comerciales, somos una empresa
-								multimarca.
-								Llámanos y realizamos una visita a tu domicilio para un
-								diagnóstico.
-							</p>
-						</div>
-						<div>
-							<h2>Área de servicio y reparación premium.</h2>
-							<img src="reparacion.webp" alt="imagene de una reparacion" />
-						</div>
-
-					</div>
-				</div>
-			</div>
-			<div className="todaslasmarcasrs">
-				<div className="titulo titulo-principalrs">
-					<h1>SERVICIO TÉCNICO TODAS LAS MARCAS</h1>
-				</div>
-				<div className="contenedorrs">
-					<div className="contenedor2rs servicio-tecnico-todasrs">
-						<div>
-							<img src="marcas/aceb.png" alt="" />
-							<img src="marcas/centrales.png" alt="" />
-							<img src="marcas/challenger.png" alt="" />
-							<img src="marcas/electrolux.png" alt="" />
-						</div>
-						<div>
-							<img src="marcas/lg.png" alt="" />
-							<img src="marcas/mabe.png" alt="" />
-							<img src="marcas/samsung.png" alt="" />
-							<img src="marcas/whirlpool.png" alt="" />
-						</div>
-					</div>
-				</div>
-			</div>
-		</section>
-	</div>
-}
-
-export default Reparacion;
