+++ react-front/src/pages/Reserva/Reserva.css
@@ -1,0 +1,51 @@
+.container .container1 p {
+    color: #000000;
+    font-family: "Roboto", sans-serif;
+  }
+  
+  .contenedor1 {
+    font-size: 25px;
+    flex: 0 0 50%;
+    padding: 20px;
+    box-sizing: border-box;
+    margin-top: 2%;
+    margin-right: 0%;
+    text-align: center;
+  }
+  
+  .contenedor2 {
+    font-size: 25px;
+    background-color: #67e8f9;
+    border: 2px solid black;
+    border-radius: 7px;
+    flex: 0 0 48%;
+    padding: 20px;
+    margin: 4px;
+    box-sizing: border-box;
+  }
+  
+  .contenedor3 {
+    font-size: 25px;
+    background-color: #67e8f9;
+    border: 2px solid black;
+    border-radius: 7px;
+    flex: 0 0 48%;
+    padding: 20px;
+    margin: 4px;
+    box-sizing: border-box;
+  }
+  
+  .lavadora {
+    flex: 0 0 50%;
+    max-width: 40%;
+    height: 40%;
+    margin: 5%;
+  }
+  
+  .container .container1 {
+    display: flex;
+    flex-wrap: wrap;
+    margin-left: auto;
+
+  }
+
