+++ react-front/src/pages/Reserva/Reserva.js
@@ -1,0 +1,33 @@
+import './Reserva.css'
+
+const AlquilaYa = () => {
+    return (
+        <div className="container container1">
+
+            <div className="contenedor1">
+                <h1>¿Por qué lavar con nosotros?</h1>
+
+                <p> Entendemos lo ocupada que puede ser la vida diaria. Es por eso que hemos diseñado nuestro servicio de alquiler de lavadoras para que sea increíblemente conveniente. Puedes reservar una lavadora según tu horario y cercanía. Lava tu ropa cuando te convenga.</p>
+            </div>
+
+            <img className="lavadora" alt="Lavadora" src="https://st3.depositphotos.com/1007566/18759/v/450/depositphotos_187597462-stock-illustration-laundry-room-with-washing-machine.jpg"/>
+
+            <div className="contenedor2">
+                <h2>Flexibilidad Horaria:</h2>
+
+                <p>¿Tienes una carga de ropa urgente a la medianoche o temprano en la mañana? No hay problema. CleanWash te ofrece la flexibilidad de acceder a nuestras lavadoras las 24 horas del día, los 7 días de la semana. Nuestro servicio está disponible cuando lo necesitas, sin restricciones de tiempo.</p>
+            </div>
+
+            <div className="contenedor3">
+                <h2>Experiencia de Usuario sin Problemas: </h2>
+
+                <p>Desde el proceso de reserva hasta la finalización del lavado, nos esforzamos por brindarte una experiencia de usuario sin problemas. Nuestra interfaz intuitiva te guiará a través del proceso de reserva, y nuestro equipo de soporte estará disponible para ayudarte en caso de cualquier pregunta o problema.</p>
+
+            </div>
+
+        </div>
+
+    );
+};
+
+export default AlquilaYa;
